Evitar la eliminación del propio usuario.

+++ TPFinal_POO2017/src/main/java/ar/edu/unobba/poo2017/tpfinal_poo2017/controller/UsuarioBacking.java
@@ -5,13 +5,17 @@
  */
 package ar.edu.unobba.poo2017.tpfinal_poo2017.controller;
 
+import ar.edu.unnoba.poo2017.tpfinal_poo2017.bundle.MessagesProducer;
 import ar.edu.unnoba.poo2017.tpfinal_poo2017.dao.UsuarioDao;
 import ar.edu.unnoba.poo2017.tpfinal_poo2017.model.Usuario;
 import java.io.Serializable;
 import java.util.List;
 import javax.annotation.PostConstruct;
 import javax.ejb.EJB;
+import javax.faces.application.FacesMessage;
+import javax.faces.context.FacesContext;
 import javax.faces.view.ViewScoped;
+import javax.inject.Inject;
 import javax.inject.Named;
 
 /**
@@ -21,41 +25,70 @@ import javax.inject.Named;
 @Named
 @ViewScoped
 public class UsuarioBacking implements Serializable {
+
     private Usuario usuario;
 
-    
+    @Inject
+    private SessionBacking session;
+
+    @Inject
+    private MessagesProducer msg;
+
     @PostConstruct
-    public void init(){
+    public void init() {
         this.usuario = new Usuario();
     }
-    
+
     @EJB
     private UsuarioDao usuarioDao;
-    
-    public List<Usuario> getUsuarios(){
+
+    public List<Usuario> getUsuarios() {
         return usuarioDao.getUsuarios();
     }
-    
-    public String create(){
-        try{
+
+    public String create() {
+        try {
             usuarioDao.create(usuario);
             return "/usuarios/index?faces-redirect=true";
-        }catch(Exception e){
+        } catch (Exception e) {
             return null;
         }
     }
-    
-    public String update(){
-        try{
+
+    public String update() {
+        try {
             usuarioDao.update(usuario);
             return "/usuarios/index?faces-redirect=true";
-        }catch(Exception e){
+        } catch (Exception e) {
             return null;
         }
     }
-    
-    public void delete(Usuario usuario){
-        usuarioDao.delete(usuario);
+
+    public void delete(Usuario usuario) {
+        FacesContext context = FacesContext.getCurrentInstance();
+        FacesMessage message;
+        if (usuario.getUsername().equals(session.getUsuario().getUsername())) {
+            message = new FacesMessage(msg.getString("usuarios_eliminarASiMismo"));
+        } else {
+            message = new FacesMessage(msg.getString("usuarios_eliminarUsuarioEliminado", usuario.getUsername()));
+            usuarioDao.delete(usuario);
+
+        }
+        context.addMessage("msgUsuario", message);
+    }
+
+    public String delete1(Usuario usuario) {
+        if (!usuario.equals(session.getUsuario())) {
+            usuarioDao.delete(usuario);
+            FacesContext context = FacesContext.getCurrentInstance();
+            FacesMessage message = new FacesMessage(msg.getString("usuarios_eliminarASiMismo"));
+            context.addMessage(null, message);
+
+            return null;
+        } else {
+            return null;
+        }
+
     }
 
     public Usuario getUsuario() {
@@ -65,6 +98,5 @@ public class UsuarioBacking implements Serializable {
     public void setUsuario(Usuario usuario) {
         this.usuario = usuario;
     }
-    
-    
+
 }
